requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
protobuf==4.25.1
pycryptodome==3.20.0
httpx==0.27.0
cachetools==5.3.3
asgiref==3.8.1
gunicorn==21.2.0
